改进数据同步工程

+++ bboss-elasticsearch-rest-file2ftp/src/main/java/org/frameworkset/tran/output/fileftp/FileUtil.java
@@ -91,26 +91,27 @@ public class FileUtil {
 			if(bw != null)
 				bw.flush();
 			this.close();
-			if(file.length() <= 0) {
-				if (fileFtpOupputContext.transferEmptyFiles()) {
+			if(!fileFtpOupputContext.disableftp()) {
+				if (file.length() <= 0) {
+					if (fileFtpOupputContext.transferEmptyFiles()) {
+						if (fileFtpOupputContext.getTransferProtocol() == FileFtpOupputContext.TRANSFER_PROTOCOL_FTP) {
+							FtpTransfer.sendFile(fileFtpOupputContext, filePath, remoteFilePath);
+						} else {
+							SFTPTransfer.sendFile(fileFtpOupputContext, this.filePath);
+						}
+					}
+				} else {
 					if (fileFtpOupputContext.getTransferProtocol() == FileFtpOupputContext.TRANSFER_PROTOCOL_FTP) {
 						FtpTransfer.sendFile(fileFtpOupputContext, filePath, remoteFilePath);
 					} else {
 						SFTPTransfer.sendFile(fileFtpOupputContext, this.filePath);
 					}
 				}
+				if (fileFtpOupputContext.backupSuccessFiles())
+					com.frameworkset.util.FileUtil.renameFile(filePath, transferSuccessFileDir);//如果文件发送成功，将文件移除到成功目录，保留一天，过期自动清理
+				else
+					com.frameworkset.util.FileUtil.deleteFile(filePath);
 			}
-			else{
-				if (fileFtpOupputContext.getTransferProtocol() == FileFtpOupputContext.TRANSFER_PROTOCOL_FTP) {
-					FtpTransfer.sendFile(fileFtpOupputContext, filePath, remoteFilePath);
-				} else {
-					SFTPTransfer.sendFile(fileFtpOupputContext, this.filePath);
-				}
-			}
-			if(fileFtpOupputContext.backupSuccessFiles())
-				com.frameworkset.util.FileUtil.renameFile(filePath,transferSuccessFileDir);//如果文件发送成功，将文件移除到成功目录，保留一天，过期自动清理
-			else
-				com.frameworkset.util.FileUtil.deleteFile(filePath);
 		}
 		catch (IOException e){
 			//com.frameworkset.util.FileUtil.renameFile(filePath,transferFailedFileDir);//如果文件发送失败，将文件移除到失败目录，定时重发
